feature: 검증 및 기능 구현
